Change int to uint in some functions

+++ hlp.go
@@ -541,8 +541,8 @@ func FileKeysInfo(dirpath string, limit int64, offset int64) (ikeys []KeysInfo, 
 
 		ik.Key = filekeys[v]
 		ik.Type = 0
-		ik.Size = infile.Size()
-		ik.Date = int32(infile.ModTime().Unix())
+		ik.Size = uint64(infile.Size())
+		ik.Date = uint32(infile.ModTime().Unix())
 
 		ikeys = append(ikeys, ik)
 
@@ -671,7 +671,7 @@ func DBKeysInfo(db *bolt.DB, ibucket string, limit int64, offset int64) (ikeys [
 
 				val := b.Get([]byte(dbkeys[v]))
 				if val != nil {
-					ik.Size = int64(Endian.Uint64(val))
+					ik.Size = Endian.Uint64(val)
 				}
 
 			} else {
@@ -694,7 +694,7 @@ func DBKeysInfo(db *bolt.DB, ibucket string, limit int64, offset int64) (ikeys [
 
 				val := b.Get([]byte(dbkeys[v]))
 				if val != nil {
-					ik.Date = int32(Endian.Uint32(val))
+					ik.Date = Endian.Uint32(val)
 				}
 
 			} else {
@@ -898,8 +898,8 @@ func AllKeysInfo(db *bolt.DB, ibucket string, dirpath string, uniq bool, limit i
 
 		ik.Key = filekeys[v]
 		ik.Type = 0
-		ik.Size = infile.Size()
-		ik.Date = int32(infile.ModTime().Unix())
+		ik.Size = uint64(infile.Size())
+		ik.Date = uint32(infile.ModTime().Unix())
 
 		ikeys = append(ikeys, ik)
 
@@ -964,7 +964,7 @@ func AllKeysInfo(db *bolt.DB, ibucket string, dirpath string, uniq bool, limit i
 
 						val := b.Get([]byte(dbkeys[v]))
 						if val != nil {
-							ik.Size = int64(Endian.Uint64(val))
+							ik.Size = Endian.Uint64(val)
 						}
 
 					} else {
@@ -987,7 +987,7 @@ func AllKeysInfo(db *bolt.DB, ibucket string, dirpath string, uniq bool, limit i
 
 						val := b.Get([]byte(dbkeys[v]))
 						if val != nil {
-							ik.Date = int32(Endian.Uint32(val))
+							ik.Date = Endian.Uint32(val)
 						}
 
 					} else {
@@ -1023,7 +1023,7 @@ func AllKeysInfo(db *bolt.DB, ibucket string, dirpath string, uniq bool, limit i
 
 					val := b.Get([]byte(dbkeys[v]))
 					if val != nil {
-						ik.Size = int64(Endian.Uint64(val))
+						ik.Size = Endian.Uint64(val)
 					}
 
 				} else {
@@ -1046,7 +1046,7 @@ func AllKeysInfo(db *bolt.DB, ibucket string, dirpath string, uniq bool, limit i
 
 					val := b.Get([]byte(dbkeys[v]))
 					if val != nil {
-						ik.Date = int32(Endian.Uint32(val))
+						ik.Date = Endian.Uint32(val)
 					}
 
 				} else {
